fix(app): preserve lazy session navigation
