+++ packages/app/src/index.ts
@@ -28,3 +28,4 @@ export {
 } from "./wsl/types"
 export { ServerConnection } from "./context/server"
 export { createDraftStore, type DraftStore } from "./utils/draft-store"
+export { preloadSessionRoute } from "./pages/session-lazy"
