+++ packages/app/src/pages/home.tsx
@@ -1,4 +1,5 @@
 import { ScrollView } from "@opencode-ai/ui/scroll-view"
+import { onCleanup, onMount } from "solid-js"
 import { createHomeController } from "./home/home-controller"
 import { createHomeProjectsController } from "./home/home-projects-controller"
 import { HomeUtilityNav } from "./home/home-projects-view"
@@ -7,8 +8,23 @@ import { createHomeScrollController } from "./home/home-scroll-controller"
 import { createHomeSessionSearchController } from "./home/home-session-search-controller"
 import { createHomeSessionsController } from "./home/home-sessions-controller"
 import { HomeSessions } from "./home/home-sessions"
+import { preloadSessionRoute } from "./session-lazy"
 
 export function Home() {
+  onMount(() => {
+    let idle: number | undefined
+    const timer = setTimeout(() => {
+      if ("requestIdleCallback" in window) {
+        idle = requestIdleCallback(() => void preloadSessionRoute(), { timeout: 3_000 })
+        return
+      }
+      void preloadSessionRoute()
+    }, 1_500)
+    onCleanup(() => {
+      clearTimeout(timer)
+      if (idle !== undefined) cancelIdleCallback(idle)
+    })
+  })
   const home = createHomeController()
   const projects = createHomeProjectsController(home)
   const sessions = createHomeSessionsController(home)
