+++ packages/app/src/pages/session-lazy.ts
@@ -1,0 +1,4 @@
+import { lazy } from "solid-js"
+
+export const TargetSessionRoute = lazy(() => import("./target-session-route"))
+export const preloadSessionRoute = TargetSessionRoute.preload
